test: split billing wallet repair tests
Changed:
Split billing wallet lifecycle repair tests into a focused repair module and moved shared wallet lifecycle setup into local test helper modules.

Benefit:
Billing wallet lifecycle coverage is easier to review and maintain without changing production behavior.
