+++ src/api/tests/service/billing/wallet_lifecycle_app_fixture.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from collections.abc import Generator
+
+from flask import Flask
+import pytest
+
+import flaskr.dao as dao
+
+
+@pytest.fixture
+def billing_wallet_lifecycle_app() -> Generator[Flask, None, None]:
+    app = Flask(__name__)
+    app.testing = True
+    app.config.update(
+        SQLALCHEMY_DATABASE_URI="sqlite:///:memory:",
+        SQLALCHEMY_BINDS={
+            "ai_shifu_saas": "sqlite:///:memory:",
+            "ai_shifu_admin": "sqlite:///:memory:",
+        },
+        SQLALCHEMY_TRACK_MODIFICATIONS=False,
+        TZ="UTC",
+    )
+    dao.db.init_app(app)
+    with app.app_context():
+        dao.db.create_all()
+        yield app
+        dao.db.session.remove()
+        dao.db.drop_all()
